refactor(constants): final cleanup of old constants imports
Remove all re-exports from utilities/constants/__init__.py except Images.
Migrate remaining callers to direct submodule imports, move NODE_STR from
aaq.py to cluster.py, and update docs and analyzer test fixtures.
