+++ docs/ARCHITECTURE_SUPPORT.md
@@ -28,13 +28,13 @@ pytest -m s390x ...
 Note: to run on the default architecture `amd64`, there's no need to set any architecture-specific markers.
 
 ## Adding new images or new architecture support
-Images for different architectures are managed under [constants.py](../utilities/constants.py) - `ArchImages`
-The data structures are defined under [images.py](../libs/infra/images.py)
+Images for different architectures are managed under [images.py](../utilities/constants/images.py) — `ArchImages`.
+The data structures are defined under [images.py](../libs/infra/images.py).
 
 ### Adding new images
 To add a new image:
 - Add the image name under the relevant dataclass under [images.py](../libs/infra/images.py)
-- Add the image name to the `ArchImages` under the relevant architecture and OS under [constants.py](../utilities/constants.py)
+- Add the image name to the `ArchImages` under the relevant architecture and OS under [images.py](../utilities/constants/images.py)
 - Add the image to the image mapping under [os_utils.py](../utilities/os_utils.py); refer to existing images for the format
 
 ### Adding new architecture support
@@ -43,6 +43,6 @@ To add a new architecture:
 - Add a new pytest marker for the architecture
 - Add a new pytest global config file for the architecture under [tests/global_config_<architecture>.py](../tests/global_config_<architecture>.py)
   - The file should contain the relevant OS matrix(es); see [global_config_amd64.py](../tests/global_config_amd64.py) for an example
-- Add the architecture name as a constant under [constants.py](../utilities/constants.py)
-- Add the architecture name to the list of supported architectures under [get_test_images_arch_class](../utilities/constants.py)
-- Add the architecture name to the `ArchImages` under the relevant architecture and OS under [constants.py](../utilities/constants.py)
+- Add the architecture name as a constant under [architecture.py](../utilities/constants/architecture.py)
+- Add the architecture name to `SUPPORTED_CPU_ARCHITECTURES` under [architecture.py](../utilities/constants/architecture.py)
+- Add the architecture name to the `ArchImages` under the relevant architecture and OS under [images.py](../utilities/constants/images.py)
