+++ tests/global_config_amd64.py
@@ -1,9 +1,9 @@
 from typing import Any
 
-from utilities.constants import (
+from utilities.constants.images import OS_FLAVOR_FEDORA
+from utilities.constants.instance_types import (
     CENTOS_STREAM9_PREFERENCE,
     CENTOS_STREAM10_PREFERENCE,
-    OS_FLAVOR_FEDORA,
     RHEL8_PREFERENCE,
     RHEL9_PREFERENCE,
     RHEL10_PREFERENCE,
