+++ tests/global_config_ibm_spectrum_sc.py
@@ -3,7 +3,7 @@ from typing import Any
 import pytest_testconfig
 from ocp_resources.datavolume import DataVolume
 
-from utilities.constants import StorageClassNames
+from utilities.constants.storage import StorageClassNames
 
 global config
 global_config = pytest_testconfig.load_python(py_file="tests/global_config.py", encoding="utf-8")
