+++ tests/global_config_oci.py
@@ -3,7 +3,11 @@ from typing import Any
 import pytest_testconfig
 from ocp_resources.datavolume import DataVolume
 
-from utilities.constants import ACCESS_MODE, VOLUME_MODE, StorageClassNames
+from utilities.constants.storage import (
+    ACCESS_MODE,
+    VOLUME_MODE,
+    StorageClassNames,
+)
 
 global config
 global_config = pytest_testconfig.load_python(py_file="tests/global_config.py", encoding="utf-8")
