+++ tests/storage/hpp/test_hpp_node_placement.py
@@ -14,7 +14,7 @@ from tests.storage.hpp.utils import (
     VM_NAME,
     edit_hpp_with_node_selector,
 )
-from utilities.constants.aaq import NODE_STR
+from utilities.constants.cluster import NODE_STR
 from utilities.storage import check_disk_count_in_vm
 
 LOGGER = logging.getLogger(__name__)
